Removed graphicplugin_factory and renamed graphicsplugin.
Moved all structs from graphicsplugin_vulkan.cpp into rvr_vulkan_utilities.h.
Made minor changes to android_platform & revolver_app.
