--- FTS/src/main/cpp/graphicsplugin_vulkan.cpp
@@ -1,1406 +1,0 @@
-#include "include/pch.h"
-#include "include/common.h"
-#include "include/geometry.h"
-#include "include/graphicsplugin.h"
-#include "include/xr_linear.h"
-#include "include/vulkan_results.h"
-#include <array>
-
-extern "C" void fast_matrix_mul(float*, float*, float*);
-
-namespace {
-struct MemoryAllocator {
-    void Init(VkPhysicalDevice physicalDevice, VkDevice device) {
-        m_vkDevice = device;
-        vkGetPhysicalDeviceMemoryProperties(physicalDevice, &m_memProps);
-    }
-
-    static const VkFlags defaultFlags = VK_MEMORY_PROPERTY_HOST_VISIBLE_BIT | VK_MEMORY_PROPERTY_HOST_COHERENT_BIT;
-
-    void Allocate(VkMemoryRequirements const& memReqs, VkDeviceMemory* mem, VkFlags flags = defaultFlags,
-                  const void* pNext = nullptr) const {
-        // Search memtypes to find first index with those properties
-        for (uint32_t i = 0; i < m_memProps.memoryTypeCount; ++i) {
-            if ((memReqs.memoryTypeBits & (1 << i)) != 0u) {
-                // Type is available, does it match user properties?
-                if ((m_memProps.memoryTypes[i].propertyFlags & flags) == flags) {
-                    VkMemoryAllocateInfo memAlloc{VK_STRUCTURE_TYPE_MEMORY_ALLOCATE_INFO, pNext};
-                    memAlloc.allocationSize = memReqs.size;
-                    memAlloc.memoryTypeIndex = i;
-                    CHECK_VKCMD(vkAllocateMemory(m_vkDevice, &memAlloc, nullptr, mem));
-                    return;
-                }
-            }
-        }
-        THROW("Memory format not supported");
-    }
-
-   private:
-    VkDevice m_vkDevice{VK_NULL_HANDLE};
-    VkPhysicalDeviceMemoryProperties m_memProps{};
-};
-
-// CmdBuffer - manage VkCommandBuffer state
-struct CmdBuffer {
-#define LIST_CMDBUFFER_STATES(_) \
-    _(Undefined)                 \
-    _(Initialized)               \
-    _(Recording)                 \
-    _(Executable)                \
-    _(Executing)
-    enum class CmdBufferState {
-#define MK_ENUM(name) name,
-        LIST_CMDBUFFER_STATES(MK_ENUM)
-#undef MK_ENUM
-    };
-    CmdBufferState state{CmdBufferState::Undefined};
-    VkCommandPool pool{VK_NULL_HANDLE};
-    VkCommandBuffer buf{VK_NULL_HANDLE};
-    VkFence execFence{VK_NULL_HANDLE};
-
-    CmdBuffer() = default;
-
-    CmdBuffer(const CmdBuffer&) = delete;
-    CmdBuffer& operator=(const CmdBuffer&) = delete;
-    CmdBuffer(CmdBuffer&&) = delete;
-    CmdBuffer& operator=(CmdBuffer&&) = delete;
-
-    ~CmdBuffer() {
-        SetState(CmdBufferState::Undefined);
-        if (m_vkDevice != nullptr) {
-            if (buf != VK_NULL_HANDLE) {
-                vkFreeCommandBuffers(m_vkDevice, pool, 1, &buf);
-            }
-            if (pool != VK_NULL_HANDLE) {
-                vkDestroyCommandPool(m_vkDevice, pool, nullptr);
-            }
-            if (execFence != VK_NULL_HANDLE) {
-                vkDestroyFence(m_vkDevice, execFence, nullptr);
-            }
-        }
-        buf = VK_NULL_HANDLE;
-        pool = VK_NULL_HANDLE;
-        execFence = VK_NULL_HANDLE;
-        m_vkDevice = nullptr;
-    }
-
-    std::string StateString(CmdBufferState s) {
-        switch (s) {
-#define MK_CASE(name)          \
-    case CmdBufferState::name: \
-        return #name;
-            LIST_CMDBUFFER_STATES(MK_CASE)
-#undef MK_CASE
-        }
-        return "(Unknown)";
-    }
-
-#define CHECK_CBSTATE(s)                                                                                           \
-    do                                                                                                             \
-        if (state != (s)) {                                                                                        \
-            Log::Write(Log::Level::Error,                                                                          \
-                       std::string("Expecting state " #s " from ") + __FUNCTION__ + ", in " + StateString(state)); \
-            return false;                                                                                          \
-        }                                                                                                          \
-    while (0)
-
-    bool Init(VkDevice device, uint32_t queueFamilyIndex) {
-        CHECK_CBSTATE(CmdBufferState::Undefined);
-
-        m_vkDevice = device;
-
-        // Create a command pool to allocate our command buffer from
-        VkCommandPoolCreateInfo cmdPoolInfo{VK_STRUCTURE_TYPE_COMMAND_POOL_CREATE_INFO};
-        cmdPoolInfo.flags = VK_COMMAND_POOL_CREATE_RESET_COMMAND_BUFFER_BIT;
-        cmdPoolInfo.queueFamilyIndex = queueFamilyIndex;
-        CHECK_VKCMD(vkCreateCommandPool(m_vkDevice, &cmdPoolInfo, nullptr, &pool));
-
-        // Create the command buffer from the command pool
-        VkCommandBufferAllocateInfo cmd{VK_STRUCTURE_TYPE_COMMAND_BUFFER_ALLOCATE_INFO};
-        cmd.commandPool = pool;
-        cmd.level = VK_COMMAND_BUFFER_LEVEL_PRIMARY;
-        cmd.commandBufferCount = 1;
-        CHECK_VKCMD(vkAllocateCommandBuffers(m_vkDevice, &cmd, &buf));
-
-        VkFenceCreateInfo fenceInfo{VK_STRUCTURE_TYPE_FENCE_CREATE_INFO};
-        CHECK_VKCMD(vkCreateFence(m_vkDevice, &fenceInfo, nullptr, &execFence));
-
-        SetState(CmdBufferState::Initialized);
-        return true;
-    }
-
-    bool Begin() {
-        CHECK_CBSTATE(CmdBufferState::Initialized);
-        VkCommandBufferBeginInfo cmdBeginInfo{VK_STRUCTURE_TYPE_COMMAND_BUFFER_BEGIN_INFO};
-        CHECK_VKCMD(vkBeginCommandBuffer(buf, &cmdBeginInfo));
-        SetState(CmdBufferState::Recording);
-        return true;
-    }
-
-    bool End() {
-        CHECK_CBSTATE(CmdBufferState::Recording);
-        CHECK_VKCMD(vkEndCommandBuffer(buf));
-        SetState(CmdBufferState::Executable);
-        return true;
-    }
-
-    bool Exec(VkQueue queue) {
-        CHECK_CBSTATE(CmdBufferState::Executable);
-
-        VkSubmitInfo submitInfo{VK_STRUCTURE_TYPE_SUBMIT_INFO};
-        submitInfo.commandBufferCount = 1;
-        submitInfo.pCommandBuffers = &buf;
-        CHECK_VKCMD(vkQueueSubmit(queue, 1, &submitInfo, execFence));
-
-        SetState(CmdBufferState::Executing);
-        return true;
-    }
-
-    bool Wait() {
-        // Waiting on a not-in-flight command buffer is a no-op
-        if (state == CmdBufferState::Initialized) {
-            return true;
-        }
-
-        CHECK_CBSTATE(CmdBufferState::Executing);
-
-        const uint32_t timeoutNs = 1 * 1000 * 1000 * 1000;
-        for (int i = 0; i < 5; ++i) {
-            auto res = vkWaitForFences(m_vkDevice, 1, &execFence, VK_TRUE, timeoutNs);
-            if (res == VK_SUCCESS) {
-                // Buffer can be executed multiple times...
-                SetState(CmdBufferState::Executable);
-                return true;
-            }
-            Log::Write(Log::Level::Info, "Waiting for CmdBuffer fence timed out, retrying...");
-        }
-
-        return false;
-    }
-
-    bool Reset() {
-        if (state != CmdBufferState::Initialized) {
-            CHECK_CBSTATE(CmdBufferState::Executable);
-
-            CHECK_VKCMD(vkResetFences(m_vkDevice, 1, &execFence));
-            CHECK_VKCMD(vkResetCommandBuffer(buf, 0));
-
-            SetState(CmdBufferState::Initialized);
-        }
-
-        return true;
-    }
-
-   private:
-    VkDevice m_vkDevice{VK_NULL_HANDLE};
-
-    void SetState(CmdBufferState newState) { state = newState; }
-
-#undef CHECK_CBSTATE
-#undef LIST_CMDBUFFER_STATES
-};
-
-// ShaderProgram to hold a pair of vertex & fragment shaders
-struct ShaderProgram {
-    std::array<VkPipelineShaderStageCreateInfo, 2> shaderInfo{
-        {{VK_STRUCTURE_TYPE_PIPELINE_SHADER_STAGE_CREATE_INFO}, {VK_STRUCTURE_TYPE_PIPELINE_SHADER_STAGE_CREATE_INFO}}};
-
-    ShaderProgram() = default;
-
-    ~ShaderProgram() {
-        if (m_vkDevice != nullptr) {
-            for (auto& si : shaderInfo) {
-                if (si.module != VK_NULL_HANDLE) {
-                    vkDestroyShaderModule(m_vkDevice, shaderInfo[0].module, nullptr);
-                }
-                si.module = VK_NULL_HANDLE;
-            }
-        }
-        shaderInfo = {};
-        m_vkDevice = nullptr;
-    }
-
-    ShaderProgram(const ShaderProgram&) = delete;
-    ShaderProgram& operator=(const ShaderProgram&) = delete;
-    ShaderProgram(ShaderProgram&&) = delete;
-    ShaderProgram& operator=(ShaderProgram&&) = delete;
-
-    void LoadVertexShader(const std::vector<char>& code) { Load(0, code); }
-
-    void LoadFragmentShader(const std::vector<char>& code) { Load(1, code); }
-
-    void Init(VkDevice device) { m_vkDevice = device; }
-
-   private:
-    VkDevice m_vkDevice{VK_NULL_HANDLE};
-
-    void Load(uint32_t index, const std::vector<char>& code) {
-        VkShaderModuleCreateInfo modInfo{VK_STRUCTURE_TYPE_SHADER_MODULE_CREATE_INFO};
-
-        auto& si = shaderInfo[index];
-        si.pName = "main";
-        std::string name;
-
-        switch (index) {
-            case 0:
-                si.stage = VK_SHADER_STAGE_VERTEX_BIT;
-                name = "vertex";
-                break;
-            case 1:
-                si.stage = VK_SHADER_STAGE_FRAGMENT_BIT;
-                name = "fragment";
-                break;
-            default:
-                THROW(Fmt("Unknown code index %d", index));
-        }
-
-        modInfo.codeSize = code.size();
-        modInfo.pCode = reinterpret_cast<const uint32_t*>(code.data());
-        CHECK_MSG((modInfo.codeSize > 0) && modInfo.pCode, Fmt("Invalid %s shader ", name.c_str()));
-
-        CHECK_VKCMD(vkCreateShaderModule(m_vkDevice, &modInfo, nullptr, &si.module));
-
-        Log::Write(Log::Level::Info, Fmt("Loaded %s shader", name.c_str()));
-    }
-};
-
-// VertexBuffer base class
-struct VertexBufferBase {
-    VkBuffer idxBuf{VK_NULL_HANDLE};
-    VkDeviceMemory idxMem{VK_NULL_HANDLE};
-    VkBuffer vtxBuf{VK_NULL_HANDLE};
-    VkDeviceMemory vtxMem{VK_NULL_HANDLE};
-    VkVertexInputBindingDescription bindDesc{};
-    std::vector<VkVertexInputAttributeDescription> attrDesc{};
-    struct {
-        uint32_t idx;
-        uint32_t vtx;
-    } count = {0, 0};
-
-    VertexBufferBase() = default;
-
-    ~VertexBufferBase() {
-        if (m_vkDevice != nullptr) {
-            if (idxBuf != VK_NULL_HANDLE) {
-                vkDestroyBuffer(m_vkDevice, idxBuf, nullptr);
-            }
-            if (idxMem != VK_NULL_HANDLE) {
-                vkFreeMemory(m_vkDevice, idxMem, nullptr);
-            }
-            if (vtxBuf != VK_NULL_HANDLE) {
-                vkDestroyBuffer(m_vkDevice, vtxBuf, nullptr);
-            }
-            if (vtxMem != VK_NULL_HANDLE) {
-                vkFreeMemory(m_vkDevice, vtxMem, nullptr);
-            }
-        }
-        idxBuf = VK_NULL_HANDLE;
-        idxMem = VK_NULL_HANDLE;
-        vtxBuf = VK_NULL_HANDLE;
-        vtxMem = VK_NULL_HANDLE;
-        bindDesc = {};
-        attrDesc.clear();
-        count = {0, 0};
-        m_vkDevice = nullptr;
-    }
-
-    VertexBufferBase(const VertexBufferBase&) = delete;
-    VertexBufferBase& operator=(const VertexBufferBase&) = delete;
-    VertexBufferBase(VertexBufferBase&&) = delete;
-    VertexBufferBase& operator=(VertexBufferBase&&) = delete;
-    void Init(VkDevice device, const MemoryAllocator* memAllocator, const std::vector<VkVertexInputAttributeDescription>& attr) {
-        m_vkDevice = device;
-        m_memAllocator = memAllocator;
-        attrDesc = attr;
-    }
-
-   protected:
-    VkDevice m_vkDevice{VK_NULL_HANDLE};
-    void AllocateBufferMemory(VkBuffer buf, VkDeviceMemory* mem) const {
-        VkMemoryRequirements memReq = {};
-        vkGetBufferMemoryRequirements(m_vkDevice, buf, &memReq);
-        m_memAllocator->Allocate(memReq, mem);
-    }
-
-   private:
-    const MemoryAllocator* m_memAllocator{nullptr};
-};
-
-// VertexBuffer template to wrap the indices and vertices
-template <typename T>
-struct VertexBuffer : public VertexBufferBase {
-    bool Create(uint32_t idxCount, uint32_t vtxCount) {
-        VkBufferCreateInfo bufInfo{VK_STRUCTURE_TYPE_BUFFER_CREATE_INFO};
-        bufInfo.usage = VK_BUFFER_USAGE_INDEX_BUFFER_BIT;
-        bufInfo.size = sizeof(uint16_t) * idxCount;
-        CHECK_VKCMD(vkCreateBuffer(m_vkDevice, &bufInfo, nullptr, &idxBuf));
-        AllocateBufferMemory(idxBuf, &idxMem);
-        CHECK_VKCMD(vkBindBufferMemory(m_vkDevice, idxBuf, idxMem, 0));
-
-        bufInfo.usage = VK_BUFFER_USAGE_VERTEX_BUFFER_BIT;
-        bufInfo.size = sizeof(T) * vtxCount;
-        CHECK_VKCMD(vkCreateBuffer(m_vkDevice, &bufInfo, nullptr, &vtxBuf));
-        AllocateBufferMemory(vtxBuf, &vtxMem);
-        CHECK_VKCMD(vkBindBufferMemory(m_vkDevice, vtxBuf, vtxMem, 0));
-
-        bindDesc.binding = 0;
-        bindDesc.stride = sizeof(T);
-        bindDesc.inputRate = VK_VERTEX_INPUT_RATE_VERTEX;
-
-        count = {idxCount, vtxCount};
-
-        return true;
-    }
-
-    void UpdateIndicies(const uint16_t* data, uint32_t elements, uint32_t offset = 0) {
-        uint16_t* map = nullptr;
-        CHECK_VKCMD(vkMapMemory(m_vkDevice, idxMem, sizeof(map[0]) * offset, sizeof(map[0]) * elements, 0, (void**)&map));
-        for (size_t i = 0; i < elements; ++i) {
-            map[i] = data[i];
-        }
-        vkUnmapMemory(m_vkDevice, idxMem);
-    }
-
-    void UpdateVertices(const T* data, uint32_t elements, uint32_t offset = 0) {
-        T* map = nullptr;
-        CHECK_VKCMD(vkMapMemory(m_vkDevice, vtxMem, sizeof(map[0]) * offset, sizeof(map[0]) * elements, 0, (void**)&map));
-        for (size_t i = 0; i < elements; ++i) {
-            map[i] = data[i];
-        }
-        vkUnmapMemory(m_vkDevice, vtxMem);
-    }
-};
-
-// RenderPass wrapper
-struct RenderPass {
-    VkFormat colorFmt{};
-    VkFormat depthFmt{};
-    VkRenderPass pass{VK_NULL_HANDLE};
-
-    RenderPass() = default;
-
-    bool Create(VkDevice device, VkFormat aColorFmt, VkFormat aDepthFmt) {
-        m_vkDevice = device;
-        colorFmt = aColorFmt;
-        depthFmt = aDepthFmt;
-
-        VkSubpassDescription subpass = {};
-        subpass.pipelineBindPoint = VK_PIPELINE_BIND_POINT_GRAPHICS;
-
-        VkAttachmentReference colorRef = {0, VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL};
-        VkAttachmentReference depthRef = {1, VK_IMAGE_LAYOUT_DEPTH_STENCIL_ATTACHMENT_OPTIMAL};
-
-        std::array<VkAttachmentDescription, 2> at = {};
-
-        VkRenderPassCreateInfo rpInfo{VK_STRUCTURE_TYPE_RENDER_PASS_CREATE_INFO};
-        rpInfo.attachmentCount = 0;
-        rpInfo.pAttachments = at.data();
-        rpInfo.subpassCount = 1;
-        rpInfo.pSubpasses = &subpass;
-
-        if (colorFmt != VK_FORMAT_UNDEFINED) {
-            colorRef.attachment = rpInfo.attachmentCount++;
-
-            at[colorRef.attachment].format = colorFmt;
-            at[colorRef.attachment].samples = VK_SAMPLE_COUNT_1_BIT;
-            at[colorRef.attachment].loadOp = VK_ATTACHMENT_LOAD_OP_CLEAR;
-            at[colorRef.attachment].storeOp = VK_ATTACHMENT_STORE_OP_STORE;
-            at[colorRef.attachment].stencilLoadOp = VK_ATTACHMENT_LOAD_OP_DONT_CARE;
-            at[colorRef.attachment].stencilStoreOp = VK_ATTACHMENT_STORE_OP_DONT_CARE;
-            at[colorRef.attachment].initialLayout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-            at[colorRef.attachment].finalLayout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-
-            subpass.colorAttachmentCount = 1;
-            subpass.pColorAttachments = &colorRef;
-        }
-
-        if (depthFmt != VK_FORMAT_UNDEFINED) {
-            depthRef.attachment = rpInfo.attachmentCount++;
-
-            at[depthRef.attachment].format = depthFmt;
-            at[depthRef.attachment].samples = VK_SAMPLE_COUNT_1_BIT;
-            at[depthRef.attachment].loadOp = VK_ATTACHMENT_LOAD_OP_CLEAR;
-            at[depthRef.attachment].storeOp = VK_ATTACHMENT_STORE_OP_STORE;
-            at[depthRef.attachment].stencilLoadOp = VK_ATTACHMENT_LOAD_OP_DONT_CARE;
-            at[depthRef.attachment].stencilStoreOp = VK_ATTACHMENT_STORE_OP_DONT_CARE;
-            at[depthRef.attachment].initialLayout = VK_IMAGE_LAYOUT_DEPTH_STENCIL_ATTACHMENT_OPTIMAL;
-            at[depthRef.attachment].finalLayout = VK_IMAGE_LAYOUT_DEPTH_STENCIL_ATTACHMENT_OPTIMAL;
-
-            subpass.pDepthStencilAttachment = &depthRef;
-        }
-
-        CHECK_VKCMD(vkCreateRenderPass(m_vkDevice, &rpInfo, nullptr, &pass));
-
-        return true;
-    }
-
-    ~RenderPass() {
-        if (m_vkDevice != nullptr) {
-            if (pass != VK_NULL_HANDLE) {
-                vkDestroyRenderPass(m_vkDevice, pass, nullptr);
-            }
-        }
-        pass = VK_NULL_HANDLE;
-        m_vkDevice = nullptr;
-    }
-
-    RenderPass(const RenderPass&) = delete;
-    RenderPass& operator=(const RenderPass&) = delete;
-    RenderPass(RenderPass&&) = delete;
-    RenderPass& operator=(RenderPass&&) = delete;
-
-   private:
-    VkDevice m_vkDevice{VK_NULL_HANDLE};
-};
-
-// VkImage + framebuffer wrapper
-struct RenderTarget {
-    VkImage colorImage{VK_NULL_HANDLE};
-    VkImage depthImage{VK_NULL_HANDLE};
-    VkImageView colorView{VK_NULL_HANDLE};
-    VkImageView depthView{VK_NULL_HANDLE};
-    VkFramebuffer fb{VK_NULL_HANDLE};
-
-    RenderTarget() = default;
-
-    ~RenderTarget() {
-        if (m_vkDevice != nullptr) {
-            if (fb != VK_NULL_HANDLE) {
-                vkDestroyFramebuffer(m_vkDevice, fb, nullptr);
-            }
-            if (colorView != VK_NULL_HANDLE) {
-                vkDestroyImageView(m_vkDevice, colorView, nullptr);
-            }
-            if (depthView != VK_NULL_HANDLE) {
-                vkDestroyImageView(m_vkDevice, depthView, nullptr);
-            }
-        }
-
-        // Note we don't own color/depthImage, it will get destroyed when xrDestroySwapchain is called
-        colorImage = VK_NULL_HANDLE;
-        depthImage = VK_NULL_HANDLE;
-        colorView = VK_NULL_HANDLE;
-        depthView = VK_NULL_HANDLE;
-        fb = VK_NULL_HANDLE;
-        m_vkDevice = nullptr;
-    }
-
-    RenderTarget(RenderTarget&& other) noexcept : RenderTarget() {
-        using std::swap;
-        swap(colorImage, other.colorImage);
-        swap(depthImage, other.depthImage);
-        swap(colorView, other.colorView);
-        swap(depthView, other.depthView);
-        swap(fb, other.fb);
-        swap(m_vkDevice, other.m_vkDevice);
-    }
-    RenderTarget& operator=(RenderTarget&& other) noexcept {
-        if (&other == this) {
-            return *this;
-        }
-        // Clean up ourselves.
-        this->~RenderTarget();
-        using std::swap;
-        swap(colorImage, other.colorImage);
-        swap(depthImage, other.depthImage);
-        swap(colorView, other.colorView);
-        swap(depthView, other.depthView);
-        swap(fb, other.fb);
-        swap(m_vkDevice, other.m_vkDevice);
-        return *this;
-    }
-    void Create(VkDevice device, VkImage aColorImage, VkImage aDepthImage, VkExtent2D size, RenderPass& renderPass) {
-        m_vkDevice = device;
-
-        colorImage = aColorImage;
-        depthImage = aDepthImage;
-
-        std::array<VkImageView, 2> attachments{};
-        uint32_t attachmentCount = 0;
-
-        // Create color image view
-        if (colorImage != VK_NULL_HANDLE) {
-            VkImageViewCreateInfo colorViewInfo{VK_STRUCTURE_TYPE_IMAGE_VIEW_CREATE_INFO};
-            colorViewInfo.image = colorImage;
-            colorViewInfo.viewType = VK_IMAGE_VIEW_TYPE_2D;
-            colorViewInfo.format = renderPass.colorFmt;
-            colorViewInfo.components.r = VK_COMPONENT_SWIZZLE_R;
-            colorViewInfo.components.g = VK_COMPONENT_SWIZZLE_G;
-            colorViewInfo.components.b = VK_COMPONENT_SWIZZLE_B;
-            colorViewInfo.components.a = VK_COMPONENT_SWIZZLE_A;
-            colorViewInfo.subresourceRange.aspectMask = VK_IMAGE_ASPECT_COLOR_BIT;
-            colorViewInfo.subresourceRange.baseMipLevel = 0;
-            colorViewInfo.subresourceRange.levelCount = 1;
-            colorViewInfo.subresourceRange.baseArrayLayer = 0;
-            colorViewInfo.subresourceRange.layerCount = 1;
-            CHECK_VKCMD(vkCreateImageView(m_vkDevice, &colorViewInfo, nullptr, &colorView));
-            attachments[attachmentCount++] = colorView;
-        }
-
-        // Create depth image view
-        if (depthImage != VK_NULL_HANDLE) {
-            VkImageViewCreateInfo depthViewInfo{VK_STRUCTURE_TYPE_IMAGE_VIEW_CREATE_INFO};
-            depthViewInfo.image = depthImage;
-            depthViewInfo.viewType = VK_IMAGE_VIEW_TYPE_2D;
-            depthViewInfo.format = renderPass.depthFmt;
-            depthViewInfo.components.r = VK_COMPONENT_SWIZZLE_R;
-            depthViewInfo.components.g = VK_COMPONENT_SWIZZLE_G;
-            depthViewInfo.components.b = VK_COMPONENT_SWIZZLE_B;
-            depthViewInfo.components.a = VK_COMPONENT_SWIZZLE_A;
-            depthViewInfo.subresourceRange.aspectMask = VK_IMAGE_ASPECT_DEPTH_BIT;
-            depthViewInfo.subresourceRange.baseMipLevel = 0;
-            depthViewInfo.subresourceRange.levelCount = 1;
-            depthViewInfo.subresourceRange.baseArrayLayer = 0;
-            depthViewInfo.subresourceRange.layerCount = 1;
-            CHECK_VKCMD(vkCreateImageView(m_vkDevice, &depthViewInfo, nullptr, &depthView));
-            attachments[attachmentCount++] = depthView;
-        }
-
-        VkFramebufferCreateInfo fbInfo{VK_STRUCTURE_TYPE_FRAMEBUFFER_CREATE_INFO};
-        fbInfo.renderPass = renderPass.pass;
-        fbInfo.attachmentCount = attachmentCount;
-        fbInfo.pAttachments = attachments.data();
-        fbInfo.width = size.width;
-        fbInfo.height = size.height;
-        fbInfo.layers = 1;
-        CHECK_VKCMD(vkCreateFramebuffer(m_vkDevice, &fbInfo, nullptr, &fb));
-    }
-
-    RenderTarget(const RenderTarget&) = delete;
-    RenderTarget& operator=(const RenderTarget&) = delete;
-
-   private:
-    VkDevice m_vkDevice{VK_NULL_HANDLE};
-};
-
-// Simple vertex MVP xform & color fragment shader layout
-struct PipelineLayout {
-    VkPipelineLayout layout{VK_NULL_HANDLE};
-
-    PipelineLayout() = default;
-
-    ~PipelineLayout() {
-        if (m_vkDevice != nullptr) {
-            if (layout != VK_NULL_HANDLE) {
-                vkDestroyPipelineLayout(m_vkDevice, layout, nullptr);
-            }
-        }
-        layout = VK_NULL_HANDLE;
-        m_vkDevice = nullptr;
-    }
-
-    void Create(VkDevice device) {
-        m_vkDevice = device;
-
-        // MVP matrix is a push_constant
-        VkPushConstantRange pcr = {};
-        pcr.stageFlags = VK_SHADER_STAGE_VERTEX_BIT;
-        pcr.offset = 0;
-        pcr.size = 4 * 4 * sizeof(float);
-
-        VkPipelineLayoutCreateInfo pipelineLayoutCreateInfo{VK_STRUCTURE_TYPE_PIPELINE_LAYOUT_CREATE_INFO};
-        pipelineLayoutCreateInfo.pushConstantRangeCount = 1;
-        pipelineLayoutCreateInfo.pPushConstantRanges = &pcr;
-        CHECK_VKCMD(vkCreatePipelineLayout(m_vkDevice, &pipelineLayoutCreateInfo, nullptr, &layout));
-    }
-
-    PipelineLayout(const PipelineLayout&) = delete;
-    PipelineLayout& operator=(const PipelineLayout&) = delete;
-    PipelineLayout(PipelineLayout&&) = delete;
-    PipelineLayout& operator=(PipelineLayout&&) = delete;
-
-   private:
-    VkDevice m_vkDevice{VK_NULL_HANDLE};
-};
-
-// Pipeline wrapper for rendering pipeline state
-struct Pipeline {
-    VkPipeline pipe{VK_NULL_HANDLE};
-    VkPrimitiveTopology topology{VK_PRIMITIVE_TOPOLOGY_TRIANGLE_LIST};
-    std::vector<VkDynamicState> dynamicStateEnables;
-
-    Pipeline() = default;
-
-    void Dynamic(VkDynamicState state) { dynamicStateEnables.emplace_back(state); }
-
-    void Create(VkDevice device, VkExtent2D size, const PipelineLayout& layout, const RenderPass& rp, const ShaderProgram& sp,
-                const VertexBufferBase& vb) {
-        m_vkDevice = device;
-
-        VkPipelineDynamicStateCreateInfo dynamicState{VK_STRUCTURE_TYPE_PIPELINE_DYNAMIC_STATE_CREATE_INFO};
-        dynamicState.dynamicStateCount = (uint32_t)dynamicStateEnables.size();
-        dynamicState.pDynamicStates = dynamicStateEnables.data();
-
-        VkPipelineVertexInputStateCreateInfo vi{VK_STRUCTURE_TYPE_PIPELINE_VERTEX_INPUT_STATE_CREATE_INFO};
-        vi.vertexBindingDescriptionCount = 1;
-        vi.pVertexBindingDescriptions = &vb.bindDesc;
-        vi.vertexAttributeDescriptionCount = (uint32_t)vb.attrDesc.size();
-        vi.pVertexAttributeDescriptions = vb.attrDesc.data();
-
-        VkPipelineInputAssemblyStateCreateInfo ia{VK_STRUCTURE_TYPE_PIPELINE_INPUT_ASSEMBLY_STATE_CREATE_INFO};
-        ia.primitiveRestartEnable = VK_FALSE;
-        ia.topology = topology;
-
-        VkPipelineRasterizationStateCreateInfo rs{VK_STRUCTURE_TYPE_PIPELINE_RASTERIZATION_STATE_CREATE_INFO};
-        rs.polygonMode = VK_POLYGON_MODE_FILL;
-        rs.cullMode = VK_CULL_MODE_BACK_BIT;
-        rs.frontFace = VK_FRONT_FACE_CLOCKWISE;
-        rs.depthClampEnable = VK_FALSE;
-        rs.rasterizerDiscardEnable = VK_FALSE;
-        rs.depthBiasEnable = VK_FALSE;
-        rs.depthBiasConstantFactor = 0;
-        rs.depthBiasClamp = 0;
-        rs.depthBiasSlopeFactor = 0;
-        rs.lineWidth = 1.0f;
-
-        VkPipelineColorBlendAttachmentState attachState{};
-        attachState.blendEnable = 0;
-        attachState.srcColorBlendFactor = VK_BLEND_FACTOR_ONE;
-        attachState.dstColorBlendFactor = VK_BLEND_FACTOR_ZERO;
-        attachState.colorBlendOp = VK_BLEND_OP_ADD;
-        attachState.srcAlphaBlendFactor = VK_BLEND_FACTOR_ONE;
-        attachState.dstAlphaBlendFactor = VK_BLEND_FACTOR_ZERO;
-        attachState.alphaBlendOp = VK_BLEND_OP_ADD;
-        attachState.colorWriteMask =
-            VK_COLOR_COMPONENT_R_BIT | VK_COLOR_COMPONENT_G_BIT | VK_COLOR_COMPONENT_B_BIT | VK_COLOR_COMPONENT_A_BIT;
-
-        VkPipelineColorBlendStateCreateInfo cb{VK_STRUCTURE_TYPE_PIPELINE_COLOR_BLEND_STATE_CREATE_INFO};
-        cb.attachmentCount = 1;
-        cb.pAttachments = &attachState;
-        cb.logicOpEnable = VK_FALSE;
-        cb.logicOp = VK_LOGIC_OP_NO_OP;
-        cb.blendConstants[0] = 1.0f;
-        cb.blendConstants[1] = 1.0f;
-        cb.blendConstants[2] = 1.0f;
-        cb.blendConstants[3] = 1.0f;
-
-        VkRect2D scissor = {{0, 0}, size};
-#if defined(ORIGIN_BOTTOM_LEFT)
-        // Flipped view so origin is bottom-left like GL (requires VK_KHR_maintenance1)
-        VkViewport viewport = {0.0f, (float)size.height, (float)size.width, -(float)size.height, 0.0f, 1.0f};
-#else
-        // Will invert y after projection
-        VkViewport viewport = {0.0f, 0.0f, (float)size.width, (float)size.height, 0.0f, 1.0f};
-#endif
-        VkPipelineViewportStateCreateInfo vp{VK_STRUCTURE_TYPE_PIPELINE_VIEWPORT_STATE_CREATE_INFO};
-        vp.viewportCount = 1;
-        vp.pViewports = &viewport;
-        vp.scissorCount = 1;
-        vp.pScissors = &scissor;
-
-        VkPipelineDepthStencilStateCreateInfo ds{VK_STRUCTURE_TYPE_PIPELINE_DEPTH_STENCIL_STATE_CREATE_INFO};
-        ds.depthTestEnable = VK_TRUE;
-        ds.depthWriteEnable = VK_TRUE;
-        ds.depthCompareOp = VK_COMPARE_OP_LESS;
-        ds.depthBoundsTestEnable = VK_FALSE;
-        ds.stencilTestEnable = VK_FALSE;
-        ds.front.failOp = VK_STENCIL_OP_KEEP;
-        ds.front.passOp = VK_STENCIL_OP_KEEP;
-        ds.front.depthFailOp = VK_STENCIL_OP_KEEP;
-        ds.front.compareOp = VK_COMPARE_OP_ALWAYS;
-        ds.back = ds.front;
-        ds.minDepthBounds = 0.0f;
-        ds.maxDepthBounds = 1.0f;
-
-        VkPipelineMultisampleStateCreateInfo ms{VK_STRUCTURE_TYPE_PIPELINE_MULTISAMPLE_STATE_CREATE_INFO};
-        ms.rasterizationSamples = VK_SAMPLE_COUNT_1_BIT;
-
-        VkGraphicsPipelineCreateInfo pipeInfo{VK_STRUCTURE_TYPE_GRAPHICS_PIPELINE_CREATE_INFO};
-        pipeInfo.stageCount = (uint32_t)sp.shaderInfo.size();
-        pipeInfo.pStages = sp.shaderInfo.data();
-        pipeInfo.pVertexInputState = &vi;
-        pipeInfo.pInputAssemblyState = &ia;
-        pipeInfo.pTessellationState = nullptr;
-        pipeInfo.pViewportState = &vp;
-        pipeInfo.pRasterizationState = &rs;
-        pipeInfo.pMultisampleState = &ms;
-        pipeInfo.pDepthStencilState = &ds;
-        pipeInfo.pColorBlendState = &cb;
-        if (dynamicState.dynamicStateCount > 0) {
-            pipeInfo.pDynamicState = &dynamicState;
-        }
-        pipeInfo.layout = layout.layout;
-        pipeInfo.renderPass = rp.pass;
-        pipeInfo.subpass = 0;
-        CHECK_VKCMD(vkCreateGraphicsPipelines(m_vkDevice, VK_NULL_HANDLE, 1, &pipeInfo, nullptr, &pipe));
-    }
-
-    void Release() {
-        if (m_vkDevice != nullptr) {
-            if (pipe != VK_NULL_HANDLE) {
-                vkDestroyPipeline(m_vkDevice, pipe, nullptr);
-            }
-        }
-        pipe = VK_NULL_HANDLE;
-        m_vkDevice = nullptr;
-    }
-
-   private:
-    VkDevice m_vkDevice{VK_NULL_HANDLE};
-};
-
-struct DepthBuffer {
-    VkDeviceMemory depthMemory{VK_NULL_HANDLE};
-    VkImage depthImage{VK_NULL_HANDLE};
-
-    DepthBuffer() = default;
-
-    ~DepthBuffer() {
-        if (m_vkDevice != nullptr) {
-            if (depthImage != VK_NULL_HANDLE) {
-                vkDestroyImage(m_vkDevice, depthImage, nullptr);
-            }
-            if (depthMemory != VK_NULL_HANDLE) {
-                vkFreeMemory(m_vkDevice, depthMemory, nullptr);
-            }
-        }
-        depthImage = VK_NULL_HANDLE;
-        depthMemory = VK_NULL_HANDLE;
-        m_vkDevice = nullptr;
-    }
-
-    DepthBuffer(DepthBuffer&& other) noexcept : DepthBuffer() {
-        using std::swap;
-
-        swap(depthImage, other.depthImage);
-        swap(depthMemory, other.depthMemory);
-        swap(m_vkDevice, other.m_vkDevice);
-    }
-    DepthBuffer& operator=(DepthBuffer&& other) noexcept {
-        if (&other == this) {
-            return *this;
-        }
-        // clean up self
-        this->~DepthBuffer();
-        using std::swap;
-
-        swap(depthImage, other.depthImage);
-        swap(depthMemory, other.depthMemory);
-        swap(m_vkDevice, other.m_vkDevice);
-        return *this;
-    }
-
-    void Create(VkDevice device, MemoryAllocator* memAllocator, VkFormat depthFormat,
-                const XrSwapchainCreateInfo& swapchainCreateInfo) {
-        m_vkDevice = device;
-
-        VkExtent2D size = {swapchainCreateInfo.width, swapchainCreateInfo.height};
-
-        // Create a D32 depthbuffer
-        VkImageCreateInfo imageInfo{VK_STRUCTURE_TYPE_IMAGE_CREATE_INFO};
-        imageInfo.imageType = VK_IMAGE_TYPE_2D;
-        imageInfo.extent.width = size.width;
-        imageInfo.extent.height = size.height;
-        imageInfo.extent.depth = 1;
-        imageInfo.mipLevels = 1;
-        imageInfo.arrayLayers = 1;
-        imageInfo.format = depthFormat;
-        imageInfo.tiling = VK_IMAGE_TILING_OPTIMAL;
-        imageInfo.initialLayout = VK_IMAGE_LAYOUT_UNDEFINED;
-        imageInfo.usage = VK_IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT;
-        imageInfo.samples = (VkSampleCountFlagBits)swapchainCreateInfo.sampleCount;
-        imageInfo.sharingMode = VK_SHARING_MODE_EXCLUSIVE;
-        CHECK_VKCMD(vkCreateImage(device, &imageInfo, nullptr, &depthImage));
-
-        VkMemoryRequirements memRequirements{};
-        vkGetImageMemoryRequirements(device, depthImage, &memRequirements);
-        memAllocator->Allocate(memRequirements, &depthMemory, VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT);
-        CHECK_VKCMD(vkBindImageMemory(device, depthImage, depthMemory, 0));
-    }
-
-    void TransitionLayout(CmdBuffer* cmdBuffer, VkImageLayout newLayout) {
-        if (newLayout == m_vkLayout) {
-            return;
-        }
-
-        VkImageMemoryBarrier depthBarrier{VK_STRUCTURE_TYPE_IMAGE_MEMORY_BARRIER};
-        depthBarrier.srcAccessMask = VK_ACCESS_DEPTH_STENCIL_ATTACHMENT_WRITE_BIT;
-        depthBarrier.dstAccessMask = VK_ACCESS_DEPTH_STENCIL_ATTACHMENT_READ_BIT;
-        depthBarrier.oldLayout = m_vkLayout;
-        depthBarrier.newLayout = newLayout;
-        depthBarrier.image = depthImage;
-        depthBarrier.subresourceRange = {VK_IMAGE_ASPECT_DEPTH_BIT, 0, 1, 0, 1};
-        vkCmdPipelineBarrier(cmdBuffer->buf, VK_PIPELINE_STAGE_ALL_GRAPHICS_BIT, VK_PIPELINE_STAGE_ALL_GRAPHICS_BIT, 0, 0, nullptr,
-                             0, nullptr, 1, &depthBarrier);
-
-        m_vkLayout = newLayout;
-    }
-
-    DepthBuffer(const DepthBuffer&) = delete;
-    DepthBuffer& operator=(const DepthBuffer&) = delete;
-
-   private:
-    VkDevice m_vkDevice{VK_NULL_HANDLE};
-    VkImageLayout m_vkLayout = VK_IMAGE_LAYOUT_UNDEFINED;
-};
-
-struct SwapchainImageContext {
-    SwapchainImageContext(XrStructureType _swapchainImageType) : swapchainImageType(_swapchainImageType) {}
-
-    // A packed array of XrSwapchainImageVulkan2KHR's for xrEnumerateSwapchainImages
-    std::vector<XrSwapchainImageVulkan2KHR> swapchainImages;
-    std::vector<RenderTarget> renderTarget;
-    VkExtent2D size{};
-    DepthBuffer depthBuffer{};
-    RenderPass rp{};
-    Pipeline pipe{};
-    XrStructureType swapchainImageType;
-
-    SwapchainImageContext() = default;
-
-    std::vector<XrSwapchainImageBaseHeader*> Create(VkDevice device, MemoryAllocator* memAllocator, uint32_t capacity,
-                                                    const XrSwapchainCreateInfo& swapchainCreateInfo, const PipelineLayout& layout,
-                                                    const ShaderProgram& sp, const VertexBuffer<Geometry::Vertex>& vb) {
-        m_vkDevice = device;
-
-        size = {swapchainCreateInfo.width, swapchainCreateInfo.height};
-        VkFormat colorFormat = (VkFormat)swapchainCreateInfo.format;
-        VkFormat depthFormat = VK_FORMAT_D32_SFLOAT;
-        // XXX handle swapchainCreateInfo.sampleCount
-
-        depthBuffer.Create(m_vkDevice, memAllocator, depthFormat, swapchainCreateInfo);
-        rp.Create(m_vkDevice, colorFormat, depthFormat);
-        pipe.Create(m_vkDevice, size, layout, rp, sp, vb);
-
-        swapchainImages.resize(capacity);
-        renderTarget.resize(capacity);
-        std::vector<XrSwapchainImageBaseHeader*> bases(capacity);
-        for (uint32_t i = 0; i < capacity; ++i) {
-            swapchainImages[i] = {swapchainImageType};
-            bases[i] = reinterpret_cast<XrSwapchainImageBaseHeader*>(&swapchainImages[i]);
-        }
-
-        return bases;
-    }
-
-    uint32_t ImageIndex(const XrSwapchainImageBaseHeader* swapchainImageHeader) {
-        auto p = reinterpret_cast<const XrSwapchainImageVulkan2KHR*>(swapchainImageHeader);
-        return (uint32_t)(p - &swapchainImages[0]);
-    }
-
-    void BindRenderTarget(uint32_t index, VkRenderPassBeginInfo* renderPassBeginInfo) {
-        if (renderTarget[index].fb == VK_NULL_HANDLE) {
-            renderTarget[index].Create(m_vkDevice, swapchainImages[index].image, depthBuffer.depthImage, size, rp);
-        }
-        renderPassBeginInfo->renderPass = rp.pass;
-        renderPassBeginInfo->framebuffer = renderTarget[index].fb;
-        renderPassBeginInfo->renderArea.offset = {0, 0};
-        renderPassBeginInfo->renderArea.extent = size;
-    }
-
-   private:
-    VkDevice m_vkDevice{VK_NULL_HANDLE};
-};
-
-struct VulkanGraphicsPlugin : public IGraphicsPlugin {
-    VulkanGraphicsPlugin(const std::shared_ptr<Options>& /*unused*/, RVRAndroidPlatform& android_platform) {
-        app_ = android_platform.GetAndroidApp();
-        m_graphicsBinding.type = GetGraphicsBindingType();
-    };
-
-    std::vector<std::string> GetInstanceExtensions() const override { return {XR_KHR_VULKAN_ENABLE2_EXTENSION_NAME}; }
-
-    // Note: The output must not outlive the input - this modifies the input and returns a collection of views into that modified
-    // input!
-    std::vector<const char*> ParseExtensionString(char* names) {
-        std::vector<const char*> list;
-        while (*names != 0) {
-            list.push_back(names);
-            while (*(++names) != 0) {
-                if (*names == ' ') {
-                    *names++ = '\0';
-                    break;
-                }
-            }
-        }
-        return list;
-    }
-
-    const char* GetValidationLayerName() {
-        uint32_t layerCount;
-        vkEnumerateInstanceLayerProperties(&layerCount, nullptr);
-        std::vector<VkLayerProperties> availableLayers(layerCount);
-        vkEnumerateInstanceLayerProperties(&layerCount, availableLayers.data());
-
-        std::vector<const char*> validationLayerNames;
-        validationLayerNames.push_back("VK_LAYER_KHRONOS_validation");
-        validationLayerNames.push_back("VK_LAYER_LUNARG_standard_validation");
-
-        // Enable only one validation layer from the list above. Prefer KHRONOS.
-        for (auto& validationLayerName : validationLayerNames) {
-            for (const auto& layerProperties : availableLayers) {
-                if (0 == strcmp(validationLayerName, layerProperties.layerName)) {
-                    return validationLayerName;
-                }
-            }
-        }
-
-        return nullptr;
-    }
-
-    void InitializeDevice(XrInstance instance, XrSystemId systemId) override {
-        // Create the Vulkan device for the adapter associated with the system.
-        // Extension function must be loaded by name
-        XrGraphicsRequirementsVulkan2KHR graphicsRequirements{XR_TYPE_GRAPHICS_REQUIREMENTS_VULKAN2_KHR};
-        CHECK_XRCMD(GetVulkanGraphicsRequirements2KHR(instance, systemId, &graphicsRequirements));
-
-        VkResult err;
-
-        std::vector<const char*> layers;
-#if !defined(NDEBUG)
-        const char* const validationLayerName = GetValidationLayerName();
-        if (validationLayerName) {
-            layers.push_back(validationLayerName);
-        } else {
-            Log::Write(Log::Level::Warning, "No validation layers found in the system, skipping");
-        }
-#endif
-
-        std::vector<const char*> extensions;
-        extensions.push_back("VK_EXT_debug_report");
-#if defined(USE_MIRROR_WINDOW)
-        extensions.push_back("VK_KHR_surface");
-#if defined(VK_USE_PLATFORM_WIN32_KHR)
-        extensions.push_back("VK_KHR_win32_surface");
-#else
-#error CreateSurface not supported on this OS
-#endif  // defined(VK_USE_PLATFORM_WIN32_KHR)
-#endif  // defined(USE_MIRROR_WINDOW)
-
-        VkApplicationInfo appInfo{VK_STRUCTURE_TYPE_APPLICATION_INFO};
-        appInfo.pApplicationName = "fts";
-        appInfo.applicationVersion = 1;
-        appInfo.pEngineName = "fts";
-        appInfo.engineVersion = 1;
-        appInfo.apiVersion = VK_API_VERSION_1_0;
-
-        VkInstanceCreateInfo instInfo{VK_STRUCTURE_TYPE_INSTANCE_CREATE_INFO};
-        instInfo.pApplicationInfo = &appInfo;
-        instInfo.enabledLayerCount = (uint32_t)layers.size();
-        instInfo.ppEnabledLayerNames = layers.empty() ? nullptr : layers.data();
-        instInfo.enabledExtensionCount = (uint32_t)extensions.size();
-        instInfo.ppEnabledExtensionNames = extensions.empty() ? nullptr : extensions.data();
-
-        XrVulkanInstanceCreateInfoKHR createInfo{XR_TYPE_VULKAN_INSTANCE_CREATE_INFO_KHR};
-        createInfo.systemId = systemId;
-        createInfo.pfnGetInstanceProcAddr = &vkGetInstanceProcAddr;
-        createInfo.vulkanCreateInfo = &instInfo;
-        createInfo.vulkanAllocator = nullptr;
-        CHECK_XRCMD(CreateVulkanInstanceKHR(instance, &createInfo, &m_vkInstance, &err));
-        CHECK_VKCMD(err);
-
-        vkCreateDebugReportCallbackEXT =
-            (PFN_vkCreateDebugReportCallbackEXT)vkGetInstanceProcAddr(m_vkInstance, "vkCreateDebugReportCallbackEXT");
-        vkDestroyDebugReportCallbackEXT =
-            (PFN_vkDestroyDebugReportCallbackEXT)vkGetInstanceProcAddr(m_vkInstance, "vkDestroyDebugReportCallbackEXT");
-        VkDebugReportCallbackCreateInfoEXT debugInfo{VK_STRUCTURE_TYPE_DEBUG_REPORT_CALLBACK_CREATE_INFO_EXT};
-        debugInfo.flags = VK_DEBUG_REPORT_ERROR_BIT_EXT | VK_DEBUG_REPORT_WARNING_BIT_EXT;
-#if !defined(NDEBUG)
-        debugInfo.flags |=
-            VK_DEBUG_REPORT_INFORMATION_BIT_EXT | VK_DEBUG_REPORT_PERFORMANCE_WARNING_BIT_EXT | VK_DEBUG_REPORT_DEBUG_BIT_EXT;
-#endif
-        debugInfo.pfnCallback = debugReportThunk;
-        debugInfo.pUserData = this;
-        CHECK_VKCMD(vkCreateDebugReportCallbackEXT(m_vkInstance, &debugInfo, nullptr, &m_vkDebugReporter));
-
-        XrVulkanGraphicsDeviceGetInfoKHR deviceGetInfo{XR_TYPE_VULKAN_GRAPHICS_DEVICE_GET_INFO_KHR};
-        deviceGetInfo.systemId = systemId;
-        deviceGetInfo.vulkanInstance = m_vkInstance;
-        CHECK_XRCMD(GetVulkanGraphicsDevice2KHR(instance, &deviceGetInfo, &m_vkPhysicalDevice));
-
-        VkDeviceQueueCreateInfo queueInfo{VK_STRUCTURE_TYPE_DEVICE_QUEUE_CREATE_INFO};
-        float queuePriorities = 0;
-        queueInfo.queueCount = 1;
-        queueInfo.pQueuePriorities = &queuePriorities;
-
-        uint32_t queueFamilyCount = 0;
-        vkGetPhysicalDeviceQueueFamilyProperties(m_vkPhysicalDevice, &queueFamilyCount, nullptr);
-        std::vector<VkQueueFamilyProperties> queueFamilyProps(queueFamilyCount);
-        vkGetPhysicalDeviceQueueFamilyProperties(m_vkPhysicalDevice, &queueFamilyCount, &queueFamilyProps[0]);
-
-        for (uint32_t i = 0; i < queueFamilyCount; ++i) {
-            // Only need graphics (not presentation) for draw queue
-            if ((queueFamilyProps[i].queueFlags & VK_QUEUE_GRAPHICS_BIT) != 0u) {
-                m_queueFamilyIndex = queueInfo.queueFamilyIndex = i;
-                break;
-            }
-        }
-
-        std::vector<const char*> deviceExtensions;
-
-        VkPhysicalDeviceFeatures features{};
-        // features.samplerAnisotropy = VK_TRUE;
-
-#if defined(USE_MIRROR_WINDOW)
-        deviceExtensions.push_back(VK_KHR_SWAPCHAIN_EXTENSION_NAME);
-#endif
-
-        VkDeviceCreateInfo deviceInfo{VK_STRUCTURE_TYPE_DEVICE_CREATE_INFO};
-        deviceInfo.queueCreateInfoCount = 1;
-        deviceInfo.pQueueCreateInfos = &queueInfo;
-        deviceInfo.enabledLayerCount = 0;
-        deviceInfo.ppEnabledLayerNames = nullptr;
-        deviceInfo.enabledExtensionCount = (uint32_t)deviceExtensions.size();
-        deviceInfo.ppEnabledExtensionNames = deviceExtensions.empty() ? nullptr : deviceExtensions.data();
-        deviceInfo.pEnabledFeatures = &features;
-
-        XrVulkanDeviceCreateInfoKHR deviceCreateInfo{XR_TYPE_VULKAN_DEVICE_CREATE_INFO_KHR};
-        deviceCreateInfo.systemId = systemId;
-        deviceCreateInfo.pfnGetInstanceProcAddr = &vkGetInstanceProcAddr;
-        deviceCreateInfo.vulkanCreateInfo = &deviceInfo;
-        deviceCreateInfo.vulkanPhysicalDevice = m_vkPhysicalDevice;
-        deviceCreateInfo.vulkanAllocator = nullptr;
-        CHECK_XRCMD(CreateVulkanDeviceKHR(instance, &deviceCreateInfo, &m_vkDevice, &err));
-        CHECK_VKCMD(err);
-
-        vkGetDeviceQueue(m_vkDevice, queueInfo.queueFamilyIndex, 0, &m_vkQueue);
-
-        m_memAllocator.Init(m_vkPhysicalDevice, m_vkDevice);
-
-        InitializeResources();
-
-        m_graphicsBinding.instance = m_vkInstance;
-        m_graphicsBinding.physicalDevice = m_vkPhysicalDevice;
-        m_graphicsBinding.device = m_vkDevice;
-        m_graphicsBinding.queueFamilyIndex = queueInfo.queueFamilyIndex;
-        m_graphicsBinding.queueIndex = 0;
-    }
-
-    std::vector<char> CreateSPIRVVector(const char* asset_name) {
-        // Load in the compiled shader from the apk
-        AAsset* file = AAssetManager_open(app_->activity->assetManager,
-                                          asset_name,
-                                          AASSET_MODE_BUFFER);
-        off_t file_length = AAsset_getLength(file);
-        char* file_content = new char[file_length];
-        AAsset_read(file, file_content, file_length);
-        std::vector<char> shader_vector(file_content, file_content + file_length);
-        delete[] file_content;
-        return shader_vector;
-    }
-
-    void InitializeResources() {
-        auto fragmentSPIRV = CreateSPIRVVector("shaders/blarg.frag.spv");
-        auto vertexSPIRV = CreateSPIRVVector("shaders/blarg.vert.spv");
-
-        if (vertexSPIRV.empty()) THROW("Failed to compile vertex shader");
-        if (fragmentSPIRV.empty()) THROW("Failed to compile fragment shader");
-
-        m_shaderProgram.Init(m_vkDevice);
-        m_shaderProgram.LoadVertexShader(vertexSPIRV);
-        m_shaderProgram.LoadFragmentShader(fragmentSPIRV);
-
-        // Semaphore to block on draw complete
-        VkSemaphoreCreateInfo semInfo{VK_STRUCTURE_TYPE_SEMAPHORE_CREATE_INFO};
-        CHECK_VKCMD(vkCreateSemaphore(m_vkDevice, &semInfo, nullptr, &m_vkDrawDone));
-
-        if (!m_cmdBuffer.Init(m_vkDevice, m_queueFamilyIndex)) THROW("Failed to create command buffer");
-
-        m_pipelineLayout.Create(m_vkDevice);
-
-        static_assert(sizeof(Geometry::Vertex) == 24, "Unexpected Vertex size");
-        m_drawBuffer.Init(m_vkDevice, &m_memAllocator,
-                          {{0, 0, VK_FORMAT_R32G32B32_SFLOAT, offsetof(Geometry::Vertex, Position)},
-                           {1, 0, VK_FORMAT_R32G32B32_SFLOAT, offsetof(Geometry::Vertex, Color)}});
-        uint32_t numCubeIdicies = sizeof(Geometry::c_cubeIndices) / sizeof(Geometry::c_cubeIndices[0]);
-        uint32_t numCubeVerticies = sizeof(Geometry::c_cubeVertices) / sizeof(Geometry::c_cubeVertices[0]);
-        m_drawBuffer.Create(numCubeIdicies, numCubeVerticies);
-        m_drawBuffer.UpdateIndicies(Geometry::c_cubeIndices, numCubeIdicies, 0);
-        m_drawBuffer.UpdateVertices(Geometry::c_cubeVertices, numCubeVerticies, 0);
-    }
-
-    int64_t SelectColorSwapchainFormat(const std::vector<int64_t>& runtimeFormats) const override {
-        // List of supported color swapchain formats.
-        constexpr int64_t SupportedColorSwapchainFormats[] = {VK_FORMAT_B8G8R8A8_SRGB, VK_FORMAT_R8G8B8A8_SRGB,
-                                                              VK_FORMAT_B8G8R8A8_UNORM, VK_FORMAT_R8G8B8A8_UNORM};
-
-        auto swapchainFormatIt =
-            std::find_first_of(runtimeFormats.begin(), runtimeFormats.end(), std::begin(SupportedColorSwapchainFormats),
-                               std::end(SupportedColorSwapchainFormats));
-        if (swapchainFormatIt == runtimeFormats.end()) {
-            THROW("No runtime swapchain format supported for color swapchain");
-        }
-
-        return *swapchainFormatIt;
-    }
-
-    const XrBaseInStructure* GetGraphicsBinding() const override {
-        return reinterpret_cast<const XrBaseInStructure*>(&m_graphicsBinding);
-    }
-
-    std::vector<XrSwapchainImageBaseHeader*> AllocateSwapchainImageStructs(
-        uint32_t capacity, const XrSwapchainCreateInfo& swapchainCreateInfo) override {
-        // Allocate and initialize the buffer of image structs (must be sequential in memory for xrEnumerateSwapchainImages).
-        // Return back an array of pointers to each swapchain image struct so the consumer doesn't need to know the type/size.
-        // Keep the buffer alive by adding it into the list of buffers.
-        m_swapchainImageContexts.emplace_back(GetSwapchainImageType());
-        SwapchainImageContext& swapchainImageContext = m_swapchainImageContexts.back();
-
-        std::vector<XrSwapchainImageBaseHeader*> bases = swapchainImageContext.Create(
-            m_vkDevice, &m_memAllocator, capacity, swapchainCreateInfo, m_pipelineLayout, m_shaderProgram, m_drawBuffer);
-
-        // Map every swapchainImage base pointer to this context
-        for (auto& base : bases) {
-            m_swapchainImageContextMap[base] = &swapchainImageContext;
-        }
-
-        return bases;
-    }
-
-    void RenderView(const XrCompositionLayerProjectionView& layerView, const XrSwapchainImageBaseHeader* swapchainImage,
-                    int64_t /*swapchainFormat*/, const std::vector<Cube>& cubes) override {
-        CHECK(layerView.subImage.imageArrayIndex == 0);  // Texture arrays not supported.
-
-        auto swapchainContext = m_swapchainImageContextMap[swapchainImage];
-        uint32_t imageIndex = swapchainContext->ImageIndex(swapchainImage);
-
-        m_cmdBuffer.Reset();
-        m_cmdBuffer.Begin();
-
-        // Ensure depth is in the right layout
-        swapchainContext->depthBuffer.TransitionLayout(&m_cmdBuffer, VK_IMAGE_LAYOUT_DEPTH_STENCIL_ATTACHMENT_OPTIMAL);
-
-        // Bind and clear eye render target
-        static XrColor4f darkSlateGrey = {0.184313729f, 0.309803933f, 0.309803933f, 1.0f};
-        static std::array<VkClearValue, 2> clearValues;
-        clearValues[0].color.float32[0] = darkSlateGrey.r;
-        clearValues[0].color.float32[1] = darkSlateGrey.g;
-        clearValues[0].color.float32[2] = darkSlateGrey.b;
-        clearValues[0].color.float32[3] = darkSlateGrey.a;
-        clearValues[1].depthStencil.depth = 1.0f;
-        clearValues[1].depthStencil.stencil = 0;
-        VkRenderPassBeginInfo renderPassBeginInfo{VK_STRUCTURE_TYPE_RENDER_PASS_BEGIN_INFO};
-        renderPassBeginInfo.clearValueCount = (uint32_t)clearValues.size();
-        renderPassBeginInfo.pClearValues = clearValues.data();
-
-        swapchainContext->BindRenderTarget(imageIndex, &renderPassBeginInfo);
-
-        vkCmdBeginRenderPass(m_cmdBuffer.buf, &renderPassBeginInfo, VK_SUBPASS_CONTENTS_INLINE);
-
-        vkCmdBindPipeline(m_cmdBuffer.buf, VK_PIPELINE_BIND_POINT_GRAPHICS, swapchainContext->pipe.pipe);
-
-        // Bind index and vertex buffers
-        vkCmdBindIndexBuffer(m_cmdBuffer.buf, m_drawBuffer.idxBuf, 0, VK_INDEX_TYPE_UINT16);
-        VkDeviceSize offset = 0;
-        vkCmdBindVertexBuffers(m_cmdBuffer.buf, 0, 1, &m_drawBuffer.vtxBuf, &offset);
-
-        // Compute the view-projection transform.
-        // Note all matrixes (including OpenXR's) are column-major, right-handed.
-        const auto& pose = layerView.pose;
-        XrMatrix4x4f proj;
-        XrMatrix4x4f_CreateProjectionFov(&proj, GRAPHICS_VULKAN, layerView.fov, 0.05f, 100.0f);
-        XrMatrix4x4f toView;
-        XrVector3f scale{1.f, 1.f, 1.f};
-        XrMatrix4x4f_CreateTranslationRotationScale(&toView, &pose.position, &pose.orientation, &scale);
-        XrMatrix4x4f view;
-        XrMatrix4x4f_InvertRigidBody(&view, &toView);
-        XrMatrix4x4f vp;
-        // Leave this in here so that we can do benchmarks later. For now just use fast_matrix_mul
-        //XrMatrix4x4f_Multiply(&vp, &proj, &view);
-        fast_matrix_mul(vp.m, proj.m, view.m);
-
-        // Render each cube
-        for (const Cube& cube : cubes) {
-            // Compute the model-view-projection transform and push it.
-            XrMatrix4x4f model;
-            XrMatrix4x4f_CreateTranslationRotationScale(&model, &cube.Pose.position, &cube.Pose.orientation, &cube.Scale);
-            XrMatrix4x4f mvp;
-
-            // Leave this in here so that we can do benchmarks later. For now just use fast_matrix_mul
-            //XrMatrix4x4f_Multiply(&mvp, &vp, &model);
-            fast_matrix_mul(mvp.m, vp.m, model.m);
-            vkCmdPushConstants(m_cmdBuffer.buf, m_pipelineLayout.layout, VK_SHADER_STAGE_VERTEX_BIT, 0, sizeof(mvp.m), &mvp.m[0]);
-
-            // Draw the cube.
-            vkCmdDrawIndexed(m_cmdBuffer.buf, m_drawBuffer.count.idx, 1, 0, 0, 0);
-        }
-
-        vkCmdEndRenderPass(m_cmdBuffer.buf);
-
-        m_cmdBuffer.End();
-        m_cmdBuffer.Exec(m_vkQueue);
-        // XXX Should double-buffer the command buffers, for now just flush
-        m_cmdBuffer.Wait();
-
-#if defined(USE_MIRROR_WINDOW)
-        // Cycle the window's swapchain on the last view rendered
-        if (swapchainContext == &m_swapchainImageContexts.back()) {
-            m_swapchain.Acquire();
-            m_swapchain.Present(m_vkQueue);
-        }
-#endif
-    }
-
-    uint32_t GetSupportedSwapchainSampleCount(const XrViewConfigurationView&) override { return VK_SAMPLE_COUNT_1_BIT; }
-
-   protected:
-    XrGraphicsBindingVulkan2KHR m_graphicsBinding{XR_TYPE_GRAPHICS_BINDING_VULKAN2_KHR};
-    std::list<SwapchainImageContext> m_swapchainImageContexts;
-    std::map<const XrSwapchainImageBaseHeader*, SwapchainImageContext*> m_swapchainImageContextMap;
-
-    VkInstance m_vkInstance{VK_NULL_HANDLE};
-    VkPhysicalDevice m_vkPhysicalDevice{VK_NULL_HANDLE};
-    VkDevice m_vkDevice{VK_NULL_HANDLE};
-    uint32_t m_queueFamilyIndex = 0;
-    VkQueue m_vkQueue{VK_NULL_HANDLE};
-    VkSemaphore m_vkDrawDone{VK_NULL_HANDLE};
-
-    MemoryAllocator m_memAllocator{};
-    ShaderProgram m_shaderProgram{};
-    CmdBuffer m_cmdBuffer{};
-    PipelineLayout m_pipelineLayout{};
-    VertexBuffer<Geometry::Vertex> m_drawBuffer{};
-
-#if defined(USE_MIRROR_WINDOW)
-    Swapchain m_swapchain{};
-#endif
-
-    PFN_vkCreateDebugReportCallbackEXT vkCreateDebugReportCallbackEXT{nullptr};
-    PFN_vkDestroyDebugReportCallbackEXT vkDestroyDebugReportCallbackEXT{nullptr};
-    VkDebugReportCallbackEXT m_vkDebugReporter{VK_NULL_HANDLE};
-
-    VkBool32 debugReport(VkDebugReportFlagsEXT flags, VkDebugReportObjectTypeEXT objectType, uint64_t object, size_t /*location*/,
-                         int32_t /*messageCode*/, const char* pLayerPrefix, const char* pMessage) {
-        std::string flagNames;
-        std::string objName;
-        Log::Level level = Log::Level::Error;
-
-        if ((flags & VK_DEBUG_REPORT_DEBUG_BIT_EXT) != 0u) {
-            flagNames += "DEBUG:";
-            level = Log::Level::Verbose;
-        }
-        if ((flags & VK_DEBUG_REPORT_INFORMATION_BIT_EXT) != 0u) {
-            flagNames += "INFO:";
-            level = Log::Level::Info;
-        }
-        if ((flags & VK_DEBUG_REPORT_PERFORMANCE_WARNING_BIT_EXT) != 0u) {
-            flagNames += "PERF:";
-            level = Log::Level::Warning;
-        }
-        if ((flags & VK_DEBUG_REPORT_WARNING_BIT_EXT) != 0u) {
-            flagNames += "WARN:";
-            level = Log::Level::Warning;
-        }
-        if ((flags & VK_DEBUG_REPORT_ERROR_BIT_EXT) != 0u) {
-            flagNames += "ERROR:";
-            level = Log::Level::Error;
-        }
-
-#define LIST_OBJECT_TYPES(_) \
-    _(UNKNOWN)               \
-    _(INSTANCE)              \
-    _(PHYSICAL_DEVICE)       \
-    _(DEVICE)                \
-    _(QUEUE)                 \
-    _(SEMAPHORE)             \
-    _(COMMAND_BUFFER)        \
-    _(FENCE)                 \
-    _(DEVICE_MEMORY)         \
-    _(BUFFER)                \
-    _(IMAGE)                 \
-    _(EVENT)                 \
-    _(QUERY_POOL)            \
-    _(BUFFER_VIEW)           \
-    _(IMAGE_VIEW)            \
-    _(SHADER_MODULE)         \
-    _(PIPELINE_CACHE)        \
-    _(PIPELINE_LAYOUT)       \
-    _(RENDER_PASS)           \
-    _(PIPELINE)              \
-    _(DESCRIPTOR_SET_LAYOUT) \
-    _(SAMPLER)               \
-    _(DESCRIPTOR_POOL)       \
-    _(DESCRIPTOR_SET)        \
-    _(FRAMEBUFFER)           \
-    _(COMMAND_POOL)          \
-    _(SURFACE_KHR)           \
-    _(SWAPCHAIN_KHR)         \
-    _(DISPLAY_KHR)           \
-    _(DISPLAY_MODE_KHR)
-
-        switch (objectType) {
-            default:
-#define MK_OBJECT_TYPE_CASE(name)                  \
-    case VK_DEBUG_REPORT_OBJECT_TYPE_##name##_EXT: \
-        objName = #name;                           \
-        break;
-                LIST_OBJECT_TYPES(MK_OBJECT_TYPE_CASE)
-
-#if VK_HEADER_VERSION >= 46
-                MK_OBJECT_TYPE_CASE(DESCRIPTOR_UPDATE_TEMPLATE_KHR)
-#endif
-#if VK_HEADER_VERSION >= 70
-                MK_OBJECT_TYPE_CASE(DEBUG_REPORT_CALLBACK_EXT)
-#endif
-        }
-
-        if ((objectType == VK_DEBUG_REPORT_OBJECT_TYPE_INSTANCE_EXT) && (strcmp(pLayerPrefix, "Loader Message") == 0) &&
-            (strncmp(pMessage, "Device Extension:", 17) == 0)) {
-            return VK_FALSE;
-        }
-
-        Log::Write(level, Fmt("%s (%s 0x%llx) [%s] %s", flagNames.c_str(), objName.c_str(), object, pLayerPrefix, pMessage));
-        if ((flags & VK_DEBUG_REPORT_ERROR_BIT_EXT) != 0u) {
-            return VK_FALSE;
-        }
-        if ((flags & VK_DEBUG_REPORT_WARNING_BIT_EXT) != 0u) {
-            return VK_FALSE;
-        }
-        return VK_FALSE;
-    }
-
-    static VKAPI_ATTR VkBool32 VKAPI_CALL debugReportThunk(VkDebugReportFlagsEXT flags, VkDebugReportObjectTypeEXT objectType,
-                                                           uint64_t object, size_t location, int32_t messageCode,
-                                                           const char* pLayerPrefix, const char* pMessage, void* pUserData) {
-        return static_cast<VulkanGraphicsPlugin*>(pUserData)->debugReport(flags, objectType, object, location, messageCode,
-                                                                          pLayerPrefix, pMessage);
-    }
-
-    virtual XrStructureType GetGraphicsBindingType() const { return XR_TYPE_GRAPHICS_BINDING_VULKAN2_KHR; }
-    virtual XrStructureType GetSwapchainImageType() const { return XR_TYPE_SWAPCHAIN_IMAGE_VULKAN2_KHR; }
-
-    virtual XrResult CreateVulkanInstanceKHR(XrInstance instance, const XrVulkanInstanceCreateInfoKHR* createInfo,
-                                             VkInstance* vulkanInstance, VkResult* vulkanResult) {
-        PFN_xrCreateVulkanInstanceKHR pfnCreateVulkanInstanceKHR = nullptr;
-        CHECK_XRCMD(xrGetInstanceProcAddr(instance, "xrCreateVulkanInstanceKHR",
-                                          reinterpret_cast<PFN_xrVoidFunction*>(&pfnCreateVulkanInstanceKHR)));
-
-        return pfnCreateVulkanInstanceKHR(instance, createInfo, vulkanInstance, vulkanResult);
-    }
-
-    virtual XrResult CreateVulkanDeviceKHR(XrInstance instance, const XrVulkanDeviceCreateInfoKHR* createInfo,
-                                           VkDevice* vulkanDevice, VkResult* vulkanResult) {
-        PFN_xrCreateVulkanDeviceKHR pfnCreateVulkanDeviceKHR = nullptr;
-        CHECK_XRCMD(xrGetInstanceProcAddr(instance, "xrCreateVulkanDeviceKHR",
-                                          reinterpret_cast<PFN_xrVoidFunction*>(&pfnCreateVulkanDeviceKHR)));
-
-        return pfnCreateVulkanDeviceKHR(instance, createInfo, vulkanDevice, vulkanResult);
-    }
-
-    virtual XrResult GetVulkanGraphicsDevice2KHR(XrInstance instance, const XrVulkanGraphicsDeviceGetInfoKHR* getInfo,
-                                                 VkPhysicalDevice* vulkanPhysicalDevice) {
-        PFN_xrGetVulkanGraphicsDevice2KHR pfnGetVulkanGraphicsDevice2KHR = nullptr;
-        CHECK_XRCMD(xrGetInstanceProcAddr(instance, "xrGetVulkanGraphicsDevice2KHR",
-                                          reinterpret_cast<PFN_xrVoidFunction*>(&pfnGetVulkanGraphicsDevice2KHR)));
-
-        return pfnGetVulkanGraphicsDevice2KHR(instance, getInfo, vulkanPhysicalDevice);
-    }
-
-    virtual XrResult GetVulkanGraphicsRequirements2KHR(XrInstance instance, XrSystemId systemId,
-                                                       XrGraphicsRequirementsVulkan2KHR* graphicsRequirements) {
-        PFN_xrGetVulkanGraphicsRequirements2KHR pfnGetVulkanGraphicsRequirements2KHR = nullptr;
-        CHECK_XRCMD(xrGetInstanceProcAddr(instance, "xrGetVulkanGraphicsRequirements2KHR",
-                                          reinterpret_cast<PFN_xrVoidFunction*>(&pfnGetVulkanGraphicsRequirements2KHR)));
-
-        return pfnGetVulkanGraphicsRequirements2KHR(instance, systemId, graphicsRequirements);
-    }
-
- private:
-    const android_app* app_;
-};
-}  // namespace
-
-std::shared_ptr<IGraphicsPlugin> CreateGraphicsPlugin_Vulkan(const std::shared_ptr<Options>& options,
-                                                             RVRAndroidPlatform& android_platform) {
-    return std::make_shared<VulkanGraphicsPlugin>(options, android_platform);
-}
