+++ FTS/src/main/cpp/include/options.h
@@ -4,6 +4,8 @@
 
 #pragma once
 
+#include <string>
+
 struct Options {
     std::string GraphicsPlugin;
