+++ FTS/src/main/cpp/include/rvr_android_platform.h
@@ -15,7 +15,7 @@ public:
 
     AAssetManager* GetAndroidAssetManager();
 
-    android_app* GetAndroidApp();
+    android_app* GetAndroidApp() const;
 
 private:
     XrInstanceCreateInfoAndroidKHR instanceCreateInfoAndroid_;
